Also move cleanup and upload_remote_data
